requirements.txt
tensorflow-cpu>=2.16
progressbar2
Pillow
numpy
click
jinja2
